rename: write/post 폴더로 이동
